rainmaker_orchestrator/requirements.txt
# Rainmaker Orchestrator Dependencies
# All versions pinned for security and reproducibility
# Last updated: 2026-01-09

# Core dependencies (from original requirements.txt)
httpx==0.28.1

# Web server dependencies
# Flask 3.1.2 - Fixes CVE-2025-47278 (signing-order vulnerability)
Flask==3.1.2

# Gunicorn 23.0.0 - Fixes CVE-2024-1135 and CVE-2024-6827 (HTTP request smuggling)
gunicorn==23.0.0

# Flask dependencies (pinned for security)
Werkzeug==3.1.3
Jinja2==3.1.5
itsdangerous==2.2.0
click==8.1.8
blinker==1.9.0

# HTTP client dependencies
httpcore==1.0.7
certifi==2024.12.14
h11==0.14.0
sniffio==1.3.1
anyio==4.7.0

# Additional production dependencies
python-dotenv==1.0.1
setproctitle==1.3.4

# AI Client dependencies
openai==1.59.6
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20

# Development dependencies (testing)
# Note: Move pytest to requirements-dev.txt in production
pytest==8.3.4
